fix: moved to admin-based routing apart from some controlled pages
