+++ frontend/src/layouts/page-layout/standard-page.component.tsx
